Fix Android crash when using whatwg-fetch library
By default, the whatwg-fetch library sets the responseType to 'blob' if the device supports blob types. However, when mocking responses, the xhr.response field is populated as a string.

Netwatch typically uses the responseType to determine if it should parse the response. When set to 'blob', it mistakenly invokes the native BlobFileReader to parse this string, leading to an error."